Reduce word count in PR template
# Description

Include a summary of the change and which issue is fixed. Include relevant motivation and context. List any dependencies that are required for this change.

In this PR, I reduce the word count of the PR template. By distilling this template down to just its essential aspects, adoption by the engineering team will improve.

# How Has This Been Tested?

Describe the tests that you ran to verify your changes. Provide instructions so we can reproduce. List any relevant details for your testing configuration.

- [x] I read through my changes thrice, (configuration: eyeglasses, surface book)


# Monitoring/KPIs

List KPI metrics that should be measured to ensure the project/deployment is successful.

- [x] word count: 267->190 words


# Documentation

- [x] Requires documentation update
  - [x] Documentation written and available [here](https://github.com/thetake/.github/blob/master/pull_request_template.md)
  - [ ] Documentation to be written later
  

# Contributor Checklist:

- [x] Code has been thoroughly reviewed
- [x] Code follows style guidelines
- [x] Code contains comments throughout, esp. in hard-to-understand areas
- [x] Unit tests pass locally, show commits successfully fix bug and/or implement feature
- [x] Monitoring plan is in place to detect issues with release
- [x] Dependent changes are published in downstream modules


# Manager/Reviewer Checklist:

- [ ] I have reviewed the modified code
- [ ] I have evaluated the potential for this change to affect other systems
- [ ] I have reviewed and/or re-run the tests outlined by the contributor
- [ ] I have reviewed any new documentation
- [ ] I have reviewed the monitoring plan and am confident that issues will be detected